Se implemento la deteccion del umbral usando el algoritmo Isodata.

+++ aipatterns/src/org/ai/pattern/gui/MainFrame.java
@@ -188,7 +188,7 @@ public void filtrar(float[] matriz){
     }
 }
 
-public void verHistograma(int[] valores, int maxnum){
+public void verHistograma(long[] valores, long maxnum){
     if(histogramadlg == null){
         histogramadlg = new HistogramaDialog(this, true);
     }
